backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch
torchvision
transformers==4.35.0
Pillow==10.0.1
numpy==1.24.3
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
